Renamed TransactionAccelerateModal's file to match component name
