+++ src/components/scenes/TransactionDetailsScene.js
@@ -22,7 +22,7 @@ import type { GuiContact, GuiWallet } from '../../types/types.js'
 import * as UTILS from '../../util/utils.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { RawTextModal } from '../modals/RawTextModal.js'
-import { TransactionAccelerateModal } from '../modals/TransactionAccelerateTransactionModal.js'
+import { TransactionAccelerateModal } from '../modals/TransactionAccelerateModal.js'
 import { TransactionAdvanceDetails } from '../modals/TransactionAdvanceDetails.js'
 import { TransactionDetailsCategoryInput } from '../modals/TransactionDetailsCategoryInput.js'
 import { TransactionDetailsFiatInput } from '../modals/TransactionDetailsFiatInput.js'
